ATTENTION READ THIS FIRST
This may be confusing. You may have trouble understanding this Token and the various purposes it has.

In summary,  GX Protocol is an Existing Non-Existent Economical Plan for a Decentralized Inter-Planetary Galatic Trading Network and the very first asset is an Existing GalaticCoin Token yet to be launched that is a project of many projects and ideas for which all are generalized towards the expansion of Data and Discoveries in Space Exploration and Research. 

For a More In-depth description, you may continue to read. Updates on the GX Protocol will be posted. 

This Describes A Token Today for the Future of Tomorrow and the Non-Existence, Existence of the Plans already being developed for an
Inter-Planetary Economy and Trading network that is also self-sustainable and able to sustain a successful economy until and after the Economies Galatic "Big Bang" which is how the people of tomorrow will teach the children of tomorrow the day the Earths Economy was Combined with the Galatic Economy for the first time causing an explosive boost to the Global Economy marking the beginning of the Galatic Rush for which can be described today as the rush of Discovering New Elements, Resources and other possible unknown materials hidden within the confines of the first planet Mankind Colonizes.